fix: Only delete if remain is empty

+++ hzbus/route_spider.py
@@ -78,7 +78,10 @@ class RouteSpider:
                 queue = json.load(f)
         self._start_tasks(queue, self.get_route_details)
         #self.get_route_details(queue)
-        os.remove(filename)
+        with open(filename, 'r') as f:
+            queue = json.load(f)
+        if not len(queue):
+            os.remove(filename)
     
     def get_route_details(self, queue: list):
         while len(queue):
